# CASEProvider/public/openapi-spec.json
{
    "swagger" : "2.0",
    "info" : {
        "version" : "1.1",
        "title" : "Competencies and Academic Standards Exchange (CASE) Service OpenAPI (JSON) Definition",
        "description" : "The Competencies and Academic Standards Exchange (CASE) Service enables the exchange of data between a Competency Records Service Provider and the consumers of the associated data. This service has been described using the 1EdTech Model Driven Specification development approach - this being the Platform Specific Model (PSM) of the service.",
        "termsOfService" : "Permission is granted to all parties to use excerpts from this document as needed in producing requests for proposals. The limited permissions granted above are perpetual and will not be revoked by 1EdTech or its successors or assigns. THIS API DEFINITION IS BEING OFFERED WITHOUT ANY WARRANTY WHATSOEVER, AND IN PARTICULAR, ANY WARRANTY OF NONINFRINGEMENT IS EXPRESSLY DISCLAIMED. ANY USE OF THIS SPECIFICATION SHALL BE MADE ENTIRELY AT THE IMPLEMENTERS OWN RISK, AND NEITHER THE  CONSORTIUM, NOR ANY OF ITS MEMBERS OR SUBMITTERS, SHALL HAVE ANY LIABILITY WHATSOEVER TO ANY IMPLEMENTER OR THIRD PARTY FOR ANY DAMAGES OF ANY NATURE WHATSOEVER,  DIRECTLY OR INDIRECTLY, ARISING FROM THE USE OF THIS API DEFINITION.",
        "contact" : {
            "name" : "Vice President of Operations, 1EdTech",
            "url" : "https://www.1edtech.org",
            "email" : "operations@1edtech.org"
        },
        "license" : {
...
    "host" : "www.imsglobal.org",
    "basePath" : "/ims/case/v1p1",
    "schemes" : [ "http","https" ],
...
    "paths" : {
...
        "/CFConcepts/{sourcedId}" : {
            "get" : {
                "operationId" : "getCFConcept",
                "summary" : "The REST read request message for the getCFConcept() API call.",
                "tags" : [ "DefinitionsManager" ],
                "description" : "This is a request to the Service Provider to provide the specified Competency Framework Concept and the set of children CFConcepts as identified by the hierarchy codes.  If the identified record cannot be found then the 'unknownobject' status code must be reported.",
                "parameters" : [
                    {
                        "name" : "sourcedId",
                        "in" : "path",
                        "description" : "The UUID that identifies the Competency Framework Concept that is to be read from the service provider.",
                        "required" : true,
                        "type" : "string"
                    }
                ],
                "x-1edtech-confidentiality" : "unrestricted",
                "responses" : {
                    "200" : {
                        "description" : "This is the response when the request has been completed successfully. It is the CFConcept and child CFConcepts from the service provider.",
                        "schema" : {
                            "$ref" : "#/definitions/CFConceptSetDType"
                        }
                    },
                    "400" : {
                        "description" : "An invalid selection field was supplied and data filtering on the selection criteria was not possible i.e. 'invalid_selection_field'. This is accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "401" : {
                        "description" : "The request was not correctly authorised i.e. 'unauthorisedrequest'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "403" : {
                        "description" : "This is used to indicate that the server can be reached and process the request but refuses to take any further action i.e. 'forbidden'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "404" : {
                        "description" : "Either the supplied identifier is unknown in the Service Provider and so the object could not be changed or an invalid UUID has been supplied. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'. The statement 'Unknown Object' of 'Invalid UUID' should also be presented.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "429" : {
                        "description" : "The server is receiving too many requests i.e. 'server_busy'. Retry at a later time. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "500" : {
                        "description" : "This code should be used only if there is catastrophic error and there is not a more appropriate code i.e. 'internal_server_error'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "default" : {
                        "description" : "This is the response data payload to be supplied when the HTTP code is NOT explicitly defined. This would be accompanied by the 'codeMajor/severity' values of 'failure/error' and the appropriate 'codeMinor' value. The associated HTTP code will also be supplied.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    }
                }
            }
        },
...
        "/CFItemTypes/{sourcedId}" : {
            "get" : {
                "operationId" : "getCFItemType",
                "summary" : "The REST read request message for the getCFItemType() API call.",
                "tags" : [ "DefinitionsManager" ],
                "description" : "This is a request to the Service Provider to provide the specified Competency Framework Item Type and the set of children CFItemTypes as identified by the hierarchy codes.  If the identified record cannot be found then the 'unknownobject' status code must be reported.",
                "parameters" : [
                    {
                        "name" : "sourcedId",
                        "in" : "path",
                        "description" : "The UUID that identifies the Competency Framework ItemType that is to be read from the service provider.",
                        "required" : true,
                        "type" : "string"
                    }
                ],
                "x-1edtech-confidentiality" : "unrestricted",
                "responses" : {
                    "200" : {
                        "description" : "This is the response when the request has been completed successfully. It is the CFItemType and child CFItemTypes from the service provider.",
                        "schema" : {
                            "$ref" : "#/definitions/CFItemTypeSetDType"
                        }
                    },
                    "400" : {
                        "description" : "An invalid selection field was supplied and data filtering on the selection criteria was not possible i.e. 'invalid_selection_field'. This is accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "401" : {
                        "description" : "The request was not correctly authorised i.e. 'unauthorisedrequest'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "403" : {
                        "description" : "This is used to indicate that the server can be reached and process the request but refuses to take any further action i.e. 'forbidden'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "404" : {
                        "description" : "Either the supplied identifier is unknown in the Service Provider and so the object could not be changed or an invalid UUID has been supplied. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'. The statement 'Unknown Object' of 'Invalid UUID' should also be presented.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "429" : {
                        "description" : "The server is receiving too many requests i.e. 'server_busy'. Retry at a later time. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "500" : {
                        "description" : "This code should be used only if there is catastrophic error and there is not a more appropriate code i.e. 'internal_server_error'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "default" : {
                        "description" : "This is the response data payload to be supplied when the HTTP code is NOT explicitly defined. This would be accompanied by the 'codeMajor/severity' values of 'failure/error' and the appropriate 'codeMinor' value. The associated HTTP code will also be supplied.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    }
                }
            }
        },
...
        "/CFLicenses/{sourcedId}" : {
            "get" : {
                "operationId" : "getCFLicense",
                "summary" : "The REST read request message for the getCFLicense() API call.",
                "tags" : [ "DefinitionsManager" ],
                "description" : "This is a request to the Service Provider to provide the specified Competency Framework License.  If the identified record cannot be found then the 'unknownobject' status code must be reported.",
                "parameters" : [
                    {
                        "name" : "sourcedId",
                        "in" : "path",
                        "description" : "The UUID that identifies the Competency Framework License that is to be read from the service provider.",
                        "required" : true,
                        "type" : "string"
                    }
                ],
                "x-1edtech-confidentiality" : "unrestricted",
                "responses" : {
                    "200" : {
                        "description" : "This is the response when the request has been completed successfully. It is the CFLicense from the service provider.",
                        "schema" : {
                            "$ref" : "#/definitions/CFLicenseDType"
                        }
                    },
                    "400" : {
                        "description" : "An invalid selection field was supplied and data filtering on the selection criteria was not possible i.e. 'invalid_selection_field'. This is accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "401" : {
                        "description" : "The request was not correctly authorised i.e. 'unauthorisedrequest'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "403" : {
                        "description" : "This is used to indicate that the server can be reached and process the request but refuses to take any further action i.e. 'forbidden'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "404" : {
                        "description" : "Either the supplied identifier is unknown in the Service Provider and so the object could not be changed or an invalid UUID has been supplied. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'. The statement 'Unknown Object' of 'Invalid UUID' should also be presented.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "429" : {
                        "description" : "The server is receiving too many requests i.e. 'server_busy'. Retry at a later time. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "500" : {
                        "description" : "This code should be used only if there is catastrophic error and there is not a more appropriate code i.e. 'internal_server_error'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "default" : {
                        "description" : "This is the response data payload to be supplied when the HTTP code is NOT explicitly defined. This would be accompanied by the 'codeMajor/severity' values of 'failure/error' and the appropriate 'codeMinor' value. The associated HTTP code will also be supplied.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    }
                }
            }
        },
...
        "/CFSubjects/{sourcedId}" : {
            "get" : {
                "operationId" : "getCFSubject",
                "summary" : "The REST read request message for the getCFSubject() API call.",
                "tags" : [ "DefinitionsManager" ],
                "description" : "This is a request to the Service Provider to provide the specified Competency Framework Subject and the set of children CFSubjects as identified by the hierarchy codes.  If the identified record cannot be found then the 'unknownobject' status code must be reported.",
                "parameters" : [
                    {
                        "name" : "sourcedId",
                        "in" : "path",
                        "description" : "The UUID that identifies the Competency Framework Subject that is to be read from the service provider.",
                        "required" : true,
                        "type" : "string"
                    }
                ],
                "x-1edtech-confidentiality" : "unrestricted",
                "responses" : {
                    "200" : {
                        "description" : "This is the response when the request has been completed successfully. It is the CFSubject and child CFSubjects from the service provider.",
                        "schema" : {
                            "$ref" : "#/definitions/CFSubjectSetDType"
                        }
                    },
                    "400" : {
                        "description" : "An invalid selection field was supplied and data filtering on the selection criteria was not possible i.e. 'invalid_selection_field'. This is accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "401" : {
                        "description" : "The request was not correctly authorised i.e. 'unauthorisedrequest'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "403" : {
                        "description" : "This is used to indicate that the server can be reached and process the request but refuses to take any further action i.e. 'forbidden'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "404" : {
                        "description" : "Either the supplied identifier is unknown in the Service Provider and so the object could not be changed or an invalid UUID has been supplied. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'. The statement 'Unknown Object' of 'Invalid UUID' should also be presented.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "429" : {
                        "description" : "The server is receiving too many requests i.e. 'server_busy'. Retry at a later time. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "500" : {
                        "description" : "This code should be used only if there is catastrophic error and there is not a more appropriate code i.e. 'internal_server_error'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    },
                    "default" : {
                        "description" : "This is the response data payload to be supplied when the HTTP code is NOT explicitly defined. This would be accompanied by the 'codeMajor/severity' values of 'failure/error' and the appropriate 'codeMinor' value. The associated HTTP code will also be supplied.",
                        "schema" : {
                            "$ref" : "#/definitions/imsx_StatusInfoDType"
                        }
                    }
                }
            }
        }
    },
    "definitions" : {
...
        "CFConceptDType" : {
            "description" : "The container for the definition of a concept which is addressed by the competency framework.",
            "type" : "object",
            "properties" : {
                "identifier" : {
                    "description" : "Model Primitive Datatype = NormalizedString. An unambiguous, synthetic, globally unique identifier for the CFConcept. This is the primary way in which the exchange identification is achieved.",
                    "type" : "string",
                    "pattern" : "[0-9a-f]{8}-[0-9a-f]{4}-[1-5]{1}[0-9a-f]{3}-[8-9a-b]{1}[0-9a-f]{3}-[0-9a-f]{12}"
                },
                "uri" : {
                    "description" : "Model Primitive Datatype = AnyURI. An unambiguous reference to the CFConcept using a network-resolvable URI. ",
                    "type" : "string",
                    "format" : "uri"
                },
                "title" : {
                    "description" : "Model Primitive Datatype = NormalizedString. The title of the CFConcept.",
                    "type" : "string"
                },
                "keywords" : {
                    "description" : "Model Primitive Datatype = NormalizedString. Defined listing of keywords delimited by '|' that the concept encompasses.",
                    "type" : "string"
                },
                "hierarchyCode" : {
                    "description" : "Model Primitive Datatype = NormalizedString. A human-referenceable code designated by the publisher to identify the item in the hierarchy of the Concepts.",
                    "type" : "string"
                },
                "description" : {
                    "description" : "Model Primitive Datatype = String. A human readable description of the CFConcept.",
                    "type" : "string"
                },
                "lastChangeDateTime" : {
                    "description" : "Model Primitive Datatype = DateTime. A system generated timestamp of the most recent change to this record. This conforms to ISO 8601 dateTime definition [ISO 8601].",
                    "type" : "string",
                    "format" : "date-time"
                },
                "extensions" : {
                    "$ref" : "#/definitions/CFConceptExtensionDType"
                }
            },
            "required" : [ "identifier","uri","title","hierarchyCode","lastChangeDateTime" ],
            "x-1edtech-confidentiality" : "unrestricted",
            "additionalProperties" : false
        },
...
        "CFConceptSetDType" : {
            "description" : "The container for the set of CFConcepts supplied in the response payload. The relationship between the CFConcepts is determined by the 'hierarchyCode'. The first CFConcept is that which has been specified in the call. The other CFConcepts are the set of children as determined by their place in the 'hierarchyCode' of the CFConcept.",
            "type" : "object",
            "properties" : {
                "CFConcepts" : {
                    "description" : "The set of CFConcepts. The relationship between the CFConcepts is determined by the 'hierarchyCode'. The first CFConcept is that which has been specified in the call. The other CFConcepts are the set of children as determined by their place in the 'hierarchyCode' of the CFConcept.",
                    "type" : "array",
                    "minItems" : 1,
                    "items" : {
                        "$ref" : "#/definitions/CFConceptDType"
                    }
                }
            },
            "required" : [ "CFConcepts" ],
            "x-1edtech-confidentiality" : "unrestricted",
            "additionalProperties" : false
        },
...
        "CFItemTypeSetDType" : {
            "description" : "The container for the set of CFItemTypes supplied in the response payload. The relationship between the CFItemTypes is determined by the 'hierarchyCode'. The first CFItemType is that which has been specified in the call. The other CFItemTypes are the set of children as determined by their place in the 'hierarchyCode' of the CFItemType.",
            "type" : "object",
            "properties" : {
                "CFItemTypes" : {
                    "description" : "The set of CFItemTypes. The relationship between the CFItemTypes is determined by the 'hierarchyCode'. The first CFItemType is that which has been specified in the call. The other CFItemTypes are the set of children as determined by their place in the 'hierarchyCode' of the CFItemType.",
                    "type" : "array",
                    "minItems" : 1,
                    "items" : {
                        "$ref" : "#/definitions/CFItemTypeDType"
                    }
                }
            },
            "required" : [ "CFItemTypes" ],
            "x-1edtech-confidentiality" : "unrestricted",
            "additionalProperties" : false
        },
        "CFLicenseDType" : {
            "description" : "The container for the information about a license used within the competency framework.",
            "type" : "object",
            "properties" : {
                "identifier" : {
                    "description" : "Model Primitive Datatype = NormalizedString. An unambiguous, synthetic, globally unique identifier for the CFLicense. This is the primary way in which the exchange identification is achieved.",
                    "type" : "string",
                    "pattern" : "[0-9a-f]{8}-[0-9a-f]{4}-[1-5]{1}[0-9a-f]{3}-[8-9a-b]{1}[0-9a-f]{3}-[0-9a-f]{12}"
                },
                "uri" : {
                    "description" : "Model Primitive Datatype = AnyURI. An unambiguous reference to the CFLicense using a network-resolvable URI. ",
                    "type" : "string",
                    "format" : "uri"
                },
                "title" : {
                    "description" : "Model Primitive Datatype = NormalizedString. The title of the CFLicense.",
                    "type" : "string"
                },
                "description" : {
                    "description" : "Model Primitive Datatype = String. A human readable description of the CFLicense.",
                    "type" : "string"
                },
                "licenseText" : {
                    "description" : "Model Primitive Datatype = String. Legal license text used by the organization to convey license permissions. This may include the actual license text, or a link to a web location containing the license as a document or as text.",
                    "type" : "string"
                },
                "lastChangeDateTime" : {
                    "description" : "Model Primitive Datatype = DateTime. A system generated timestamp of the most recent change to this record. This conforms to ISO 8601 dateTime definition [ISO 8601].",
                    "type" : "string",
                    "format" : "date-time"
                },
                "extensions" : {
                    "$ref" : "#/definitions/CFLicenseExtensionDType"
                }
            },
            "required" : [ "identifier","uri","title","licenseText","lastChangeDateTime" ],
            "x-1edtech-confidentiality" : "unrestricted",
            "additionalProperties" : false
        },
        "CFLicenseExtensionDType" : {
            "description" : "The container for the proprietary extensions to the CFLicense class.",
            "type" : "object",
            "properties" : {

            },
            "x-1edtech-privacy" : [ "extensions" ],
            "x-1edtech-confidentiality" : "unrestricted",
            "additionalProperties" : true
        },
...
        "CFSubjectSetDType" : {
            "description" : "The container for the set of CFSubjects supplied in the response payload. The relationship between the CFSubjects is determined by the 'hierarchyCode'. The first CFSubject is that which has been specified in the call. The other CFSubjects are the set of children as determined by their place in the 'hierarchyCode' of the CFSubject.",
            "type" : "object",
            "properties" : {
                "CFSubjects" : {
                    "description" : "The set of CFSubjects. The relationship between the CFSubjects is determined by the 'hierarchyCode'. The first CFSubject is that which has been specified in the call. The other CFSubjects are the set of children as determined by their place in the 'hierarchyCode' of the CFSubject.",
                    "type" : "array",
                    "minItems" : 1,
                    "items" : {
                        "$ref" : "#/definitions/CFSubjectDType"
                    }
                }
            },
            "required" : [ "CFSubjects" ],
            "x-1edtech-confidentiality" : "unrestricted",
            "additionalProperties" : false
        },
...
        "imsx_StatusInfoDType" : {
            "description" : "This is the container for the status code and associated information returned within the HTTP messages received from the Service Provider. For the CASE service this object will only be returned to provide information about a failed request i.e. it will NOT be in the payload for a successful request. See Appendix B for further information on the interpretation of the information contained within this class",
            "type" : "object",
            "properties" : {
                "imsx_codeMajor" : {
                    "description" : "The code major value (from the corresponding enumerated vocabulary). See Appendix B for further information on the interpretation of this set of codes.",
                    "type" : "string",
                    "enum" : [ "failure","processing","success","unsupported" ]
                },
                "imsx_severity" : {
                    "description" : "The severity value (from the corresponding enumerated vocabulary). See Appendix B for further information on the interpretation of this set of codes.",
                    "type" : "string",
                    "enum" : [ "error","status","warning" ]
                },
                "imsx_description" : {
                    "description" : "Model Primitive Datatype = String. A human readable description supplied by the entity creating the status code information.",
                    "type" : "string"
                },
                "imsx_codeMinor" : {
                    "$ref" : "#/definitions/imsx_CodeMinorDType"
                }
            },
            "required" : [ "imsx_codeMajor","imsx_severity" ],
            "x-1edtech-confidentiality" : "unrestricted",
            "additionalProperties" : false
        }
    },
    "consumes" : [ "application/json" ],
    "produces" : [ "application/json" ]
}